refactor: Move entry points to root dir
