+++ package.json
@@ -4,10 +4,10 @@
   "description": "Micro-service used by NEARStudio to deploy smart contracts.",
   "main": "app.js",
   "scripts": {
-    "start": "env $(sed 's/ # .*//' .env) supervisor ./src/app",
-    "start:indexer": "env $(sed 's/ # .*//' .env) supervisor ./src/indexer",
-    "docker:start": "node ./src/app.js",
-    "docker:start:indexer": "node ./src/indexer.js",
+    "start": "env $(sed 's/ # .*//' .env) supervisor app",
+    "start:indexer": "env $(sed 's/ # .*//' .env) supervisor indexer",
+    "docker:start": "node app.js",
+    "docker:start:indexer": "node indexer.js",
     "test": "yarn lint && mocha",
     "lint": "eslint .",
     "fix": "eslint . --fix"
