+++ test/createTestServerInstance.js
@@ -1,6 +1,6 @@
 const supertest = require('supertest');
 
-const app = require('../src/app');
+const app = require('../app');
 
 // TODO: Inject services / service methods for route binding here rather than having them hard coded in `app.js`
 function createTestServerInstance() {
